Fix Windows build: move FIFO tests behind unix build tag
syscall.Mkfifo doesn't compile on Windows, so the runtime skip in
upload_test.go fired too late — cross-platform CI failed at build
time. Moved the two FIFO tests to upload_fifo_unix_test.go tagged
//go:build unix.
